requirements.txt
Django==4.2.7
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
mysqlclient==2.2.0
openpyxl==3.1.2
pandas==2.1.3
